+++ src/client/app/routes/index.jsx
@@ -4,11 +4,13 @@ import { Route, Switch } from 'react-router-dom';
 
 import { Home } from '../components/home';
 import { PageNotFound } from '../components/page-not-found';
+import { AboutContainer } from '../components/about';
 
 const routes = (
   <div>
     <Switch>
       <Route exact path='/' component={Home} />
+      <Route path='/about' component={AboutContainer} />
       <Route component={PageNotFound} />
     </Switch>
   </div>