+++ CHANGELOG.md
@@ -3,8 +3,13 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased]
+### Added
+- Added connected-react-router support
+- Added simple About container component to test routing
+
 ### Changed
 - Updated several dependencies to their latest versions
+- Updated reducer HMR to account for connected-react-router
 
 ## [1.3.0] - 2017-02-06
 ### Added